+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 * Infrastructure maintenance:
   * flutter pub upgrade.
   * flutter upgrade to 3.35.3.
-  * 
+
 ## 0.2.14 - 2025-09-04
 * New features:
   * Get list of available encodings from 'therion --print-encodings'.