+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * New features:
   * Loading images on web.
   * Added keyboard shortcuts for initial page.
+  * Support Cmd+keyboard shortcuts on MacOS.
 * Fixed bugs:
 * Infrastructure maintenance:
   * flutter pub upgrade.